telegram_bot/requirements.txt
python-telegram-bot==21.7
aiohttp==3.9.5
python-dotenv==1.0.1
pytz==2024.1
